Improve navigation parenting.

+++ src/views/auth/index.blade.php
@@ -25,11 +25,13 @@
                     <td>{{ $user->name }}</td>
                     <td>{{ $user->email }}</td>
                     <td>
-                        {!! Form::open(['method'=>'DELETE','action'=>['\BruceCms\Pages\AuthenticationController@destroy', $user->id], 
-                                'class' => '+inline-block']) !!}
-                        {!! Form::submit('Delete', ['class' => 'btn btn-danger', 
-                                'onclick' =>'return confirm("Are you sure? This cannot be undone.");']) !!}
-                        {!! Form::close() !!}
+                        @if($user->id !== \Auth::user()->id)
+                            {!! Form::open(['method'=>'DELETE','action'=>['\BruceCms\Pages\AuthenticationController@destroy', $user->id], 
+                                    'class' => '+inline-block']) !!}
+                            {!! Form::submit('Delete', ['class' => 'btn btn-danger', 
+                                    'onclick' =>'return confirm("Are you sure? This cannot be undone.");']) !!}
+                            {!! Form::close() !!}
+                        @endif
                     </td>
                 </tr>
             @endforeach
